+++ src/components/molecules/ReviewBtns.jsx
@@ -2,7 +2,7 @@
 
 import { jsx, css } from '@emotion/core'
 import { Link } from 'react-router-dom';
-import BtnStyle from '../atoms/BtnStyle';
+import Button from '../atoms/Button';
 import { Redirect } from 'react-router-dom';
 
 const Styles = {
@@ -20,9 +20,9 @@ const ReviewBtns = ({ movieId }) => {
   return (
     <div css={Styles.reviewBtn}>
       <Link to={'/movie/' + movieId} css={Styles.cancellink}>
-        <BtnStyle btnText="Cancel" style="cancelBtn" />
+        <Button text="Cancel" style="cancelBtn" />
       </Link>
-      <BtnStyle btnText="Publish" style="publishBtn" />
+      <Button text="Publish" style="publishBtn" />
     </div>
   )
 }
