+++ src/components/molecules/SearchBar.jsx
@@ -1,7 +1,7 @@
 /** @jsx jsx */
 
 import { jsx, css } from '@emotion/core'
-import BtnStyle from '../atoms/BtnStyle';
+import Button from '../atoms/Button';
 import SearchInput from '../atoms/SearchInput';
 
 const SearchBar = ({ handleChange }) => {
@@ -9,7 +9,7 @@ const SearchBar = ({ handleChange }) => {
   return (
     <div>
       <SearchInput style="inputDefault" handleChange={handleChange} />
-      <BtnStyle btnText="Search" />
+      <Button btnText="Search" />
     </div>
   )
 }
