+++ src/createStore.jsx
@@ -2,11 +2,9 @@ import { createStore as reduxCreateStore, applyMiddleware, combineReducers } fro
 import logger from 'redux-logger';
 import thunk from 'redux-thunk';
 import { movieReducer } from './reducers/movieReducer';
-// import { composeWithDevTools } from 'remote-redux-devtools';
 import { composeWithDevTools } from 'redux-devtools-extension';
 
-const composeEnhancers = composeWithDevTools({ realtime: true });
-// const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose
+const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose
 
 export default function createStore() {
     const store = reduxCreateStore(
